requirements.txt
tavily-python==0.7.5
langchain-community==0.3.25
pypdf==5.1.0
langchain==0.3.25
langgraph==0.4.5
langchain-openai==0.3.18
python-dotenv==1.1.0
gradio==5.34.2
langgraph-prebuilt==0.2.2
opik
fastmcp==2.8.1
langchain-mcp-adapters==0.1.7
gradio==5.34.2
huggingface_hub[cli]
